Reorganize files to make project more readable.
